Update read_metadata to use XML format.
When not specifying a format within read_metadata, the request fails with the following error: "The JSON metadata is not supported at this platform. It's only supported at platform implementing .NETStardard 2.0."

This update changes read_metadata to specify the format: :xml option on the request, which allows the metadata to be read successfully.